+++ src/App.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom'
-import Home from './assets/Components/Home/Home'
-import LogIn from './assets/Components/LogIn/LogIn'
-import SignUp from './assets/Components/SignUp/SignUp'
-import NavBar from './assets/Components/NavBar/NavBar'
-import Profile from './assets/Components/Profile/Profile'
-import { AuthProvider } from './assets/helpers/AuthContext'
-import ShowcaseForm from './assets/Components/Showcase/ShowcaseForm'
+import Home from './Components/Home/Home'
+import LogIn from './Components/LogIn/LogIn'
+import SignUp from './Components/SignUp/SignUp'
+import NavBar from './Components/NavBar/NavBar'
+import Profile from './Components/Profile/Profile'
+import { AuthProvider } from './helpers/AuthContext'
+import ShowcaseForm from './Components/Showcase/ShowcaseForm'
 
 function App() {
   return (
